Making building and leasing work
I made building and leasing work, adding the buttons back and
transacting sales using the Sales class.

For now, all lots in the "homesite" are owned by the player. This is for
testing purposes and will likely not remain into the final game. If you
want to try leasing lots, go to a different site.